refactor: move metrics to @linzjs/metrics
