+++ packages/linzjs-metrics/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@basemaps/metrics",
+  "name": "@linzjs/metrics",
   "version": "4.11.1",
   "repository": {
     "type": "git",
