+++ packages/tiler-sharp/package.json
@@ -19,7 +19,7 @@
   },
   "dependencies": {
     "@basemaps/geo": "^4.11.1",
-    "@basemaps/metrics": "^4.11.1",
+    "@linzjs/metrics": "^4.11.1",
     "@basemaps/tiler": "^4.11.1",
     "sharp": "^0.25.2"
   },
