+++ tsconfig.json
@@ -4,7 +4,7 @@
         { "path": "./packages/__tests__" },
         { "path": "./packages/linzjs-geojson" },
         { "path": "./packages/linzjs-s3fs" },
-        { "path": "./packages/metrics" },
+        { "path": "./packages/linzjs-metrics" },
         { "path": "./packages/geo" },
         { "path": "./packages/tiler" },
         { "path": "./packages/tiler-sharp" },
